Pacman game
Pacman, often stylized as "PAC-MAN," is a classic arcade video game created by the Japanese company Namco (now known as Bandai Namco Entertainment). It was first released in 1980 and quickly became one of the most iconic and recognizable video games in history. Pacman is considered a pioneering game in the arcade and video game.